+++ integrations/tailwindcss-cli/tests/cli.test.js
@@ -207,6 +207,52 @@ describe('Build command', () => {
     )
   })
 
+  test('configs support import.meta', async () => {
+    // Skip this test in Node 18 as this only works with
+    // `require(esm)` in Node 20.19+
+    if (process.versions.node.startsWith('18.')) {
+      expect(true).toBe(true)
+      return
+    }
+
+    await writeInputFile('index.html', html`<div class="font-bold"></div>`)
+
+    let customConfig = `
+      console.log(import.meta.url)
+      console.log(import.meta.resolve('./tailwind.config.mjs'))
+      export default ${JSON.stringify(
+        {
+          content: ['./src/index.html'],
+          theme: {
+            extend: {
+              fontWeight: {
+                bold: 'BOLD',
+              },
+            },
+          },
+          corePlugins: {
+            preflight: false,
+          },
+          plugins: [],
+        },
+        null,
+        2
+      )}
+    `
+
+    await writeInputFile('../tailwind.config.mjs', customConfig)
+
+    await $(`${EXECUTABLE} --output ./dist/main.css --config ./tailwind.config.mjs`)
+
+    expect(await readOutputFile('main.css')).toIncludeCss(
+      css`
+        .font-bold {
+          font-weight: BOLD;
+        }
+      `
+    )
+  })
+
   test('--content', async () => {
     await writeInputFile('other.html', html`<div class="font-bold"></div>`)
 
@@ -377,6 +423,79 @@ describe('Build command', () => {
     await writeInputFile('../custom.postcss.config.js', customConfig)
 
     await $(`${EXECUTABLE} --output ./dist/main.css --postcss ./custom.postcss.config.js`)
+
+    let contents = await readOutputFile('main.css')
+
+    expect(contents).toIncludeCss(
+      css`
+        .font-bold {
+          font-weight: 700;
+        }
+      `
+    )
+
+    expect(contents).toContain(`/*# sourceMappingURL`)
+  })
+
+  test('--postcss supports ESM configs', async () => {
+    await writeInputFile('index.html', html`<div class="font-bold"></div>`)
+
+    let customConfig = javascript`
+      import * as path from 'path'
+      import { createRequire } from 'module'
+      const require = createRequire(import.meta.url)
+
+      export default {
+        map: { inline: true },
+        plugins: [
+          function tailwindcss() {
+            return require(path.resolve('..', '..'))
+          },
+        ],
+      }
+    `
+
+    await removeFile('./postcss.config.js')
+    await writeInputFile('../postcss.config.mjs', customConfig)
+
+    await $(`${EXECUTABLE} --output ./dist/main.css --postcss`)
+
+    let contents = await readOutputFile('main.css')
+
+    expect(contents).toIncludeCss(
+      css`
+        .font-bold {
+          font-weight: 700;
+        }
+      `
+    )
+
+    expect(contents).toContain(`/*# sourceMappingURL`)
+  })
+
+  test('--postcss supports TS configs', async () => {
+    await writeInputFile('index.html', html`<div class="font-bold"></div>`)
+
+    let customConfig = javascript`
+      import * as path from 'path'
+      import { createRequire } from 'module'
+      import type { AcceptedPlugin } from 'postcss'
+      const require = createRequire(import.meta.url)
+
+      export default {
+        map: { inline: true },
+        plugins: [
+          function tailwindcss() {
+            return require(path.resolve('..', '..'))
+          } as AcceptedPlugin,
+        ],
+      }
+    `
+
+    await removeFile('./postcss.config.js')
+    await writeInputFile('../postcss.config.ts', customConfig)
+
+    await $(`${EXECUTABLE} --output ./dist/main.css --postcss`)
 
     let contents = await readOutputFile('main.css')